+++ src/value.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, Clone)]
 pub struct Value {
-    pub value: u64,
+    pub value: u16,
     pub col_sum: usize,
 }
